requirements.txt
llama_cpp_python==0.2.71
Flask==3.0.3
huggingface-hub==0.23.0